+++ src/index.ts
@@ -1,23 +1,27 @@
-import { unlink, Options, Callback} from './secure-rm'
-import { validIDs, standards, UnlinkStandard } from './standards'
-import { write, eventEmitter } from './write'
+import { remove, Options, Callback } from './secure-rm'// eslint-disable-line no-unused-vars
+import { validIDs, standards, Standard } from './standards'
+import { eventEmitter } from './events'
+import Unlink from './unlink'
+import RmDir from './rmdir'
 
 interface SecureRm {
-  (path: string, options?: Options | Callback, callback?: Callback): typeof unlink;
-  event: typeof eventEmitter;
-  write: typeof write;
-  standards: typeof standards;
-  validIDs: typeof validIDs;
-  UnlinkStandard: typeof UnlinkStandard;
-  default: SecureRm;
+  (path: string, options?: Options | Callback, callback?: Callback): typeof remove
+  event: typeof eventEmitter
+  standards: typeof standards
+  validIDs: typeof validIDs
+  Unlink: typeof Unlink
+  RmDir: typeof RmDir
+  Standard: typeof Standard
+  default: SecureRm
 }
 
-let secureRmExport = <unknown>unlink as SecureRm
+const secureRmExport = <unknown>remove as SecureRm
 secureRmExport.event = eventEmitter
-secureRmExport.write = write
 secureRmExport.standards = standards
 secureRmExport.validIDs = validIDs
-secureRmExport.UnlinkStandard = UnlinkStandard
+secureRmExport.Unlink = Unlink
+secureRmExport.RmDir = RmDir
+secureRmExport.Standard = Standard
 secureRmExport.default = secureRmExport
 
 export = secureRmExport