+++ CHANGELOG.md
@@ -12,7 +12,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security -->
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [4.0.0] - 2019-10-27
+
+### Added
+- Custom directory removal.
+- More tests.
+- Wiki.
+- Preview standard.
+
+### Changed
+- Repository location.
+- New custom standard syntax.
+
+### Security
+- Libraries update.
 
 ## [3.0.3] - 2019-08-30